refactor(buildCachedGeometryFromTree): put all relevant code into a single helper function to generate geometry from a vtree while using caching
