+++ packages/io/jscad-tree-experiments/user/index.js
@@ -3,6 +3,8 @@ const generateTree = require('../core/geometry-tree-generator')
 const findDisconnectedSubGraphs = require('../core/findDisconnectedSubGraphs')
 const writeOutput = require('../io/writeOutput')
 
+
+const makeBuildCachedGeometryFromTree = require('../core/buildCachedGeometryFromTree')
 const {runCompare, runVTreeTree} = require('./runCompare')
 /* runCompare('./examples/logo')
 runCompare('./examples/basic')
@@ -10,39 +12,25 @@ runCompare('./examples/union')
 runCompare('./examples/shapes-array') */
 
 //
-const hash = require('object-hash')
-const lookup = {}
-const buildFinalResult = (subTrees) => {
-  const finalResult = []
-  subTrees.forEach(function (indy, index) {
-    const subTreeHash = hash(indy)
-    const foundData = lookup[subTreeHash]
-    if (foundData !== undefined) {
-      // console.log('foundData', foundData.subTree)
-      finalResult.push(foundData.geom)
-    } else {
-      const subTreeGeom = generate(indy)
-      lookup[subTreeHash] = subTreeGeom
-      finalResult.push(subTreeGeom)
-    }
-  })
-  return finalResult
-}
+/*let fooVtree = require('./examples/transforms-vtree')()
+const subTrees = findDisconnectedSubGraphs(fooVtree)
+console.log('foo', JSON.stringify(subTrees))
+return */
+
+const buildCachedGeometryFromTree = makeBuildCachedGeometryFromTree()
 
 let start = process.hrtime()
 console.log('start optimised, first iteration')
-let results = require('./examples/complex-vtree')()
-const subTrees = findDisconnectedSubGraphs(results)
-buildFinalResult(subTrees)
+let vtree = require('./examples/complex-vtree')()
+let final = buildCachedGeometryFromTree(vtree)
 let elapsed = process.hrtime(start)[1] / 1000000
 
 console.log('optimised time', elapsed, 'ms')
 
 console.log('start optimised, changed one of output of main() (array of csgs)')
 start = process.hrtime()
-let resultsChanged = require('./examples/complex-vtree-changed')()
-const subTreesChanged = findDisconnectedSubGraphs(resultsChanged)
-buildFinalResult(subTreesChanged)
+vtree = require('./examples/complex-vtree-changed')()
+final = buildCachedGeometryFromTree(vtree)
 
 elapsed = process.hrtime(start)[1] / 1000000
 console.log('optimised time', elapsed, 'ms')
@@ -60,13 +48,9 @@ require('./examples/complex-base')()
 elapsed = process.hrtime(start)[1] / 1000000
 console.log('vanilla time', elapsed, 'ms')
 
-// let geomResults = generateTree(results)
-
+//
 /* writeOutput('foo.stl', generate(results))
 console.log('output written')
-
-let curTree = results
-//
 
 const {diff} = require('just-diff')
 // const diff = require('simple-diff')
